회원가입 오류 수정

+++ app/src/main/java/com/example/letsbasket/login/RegisterRequest.kt
@@ -8,8 +8,8 @@ data class RegisterRequest(
     var studentnum: Int? = null,
     var major: String? = null,
     var dong: String? = null,
-    var latitude: Float? = null,
-    var longtitude: Float? = null,
+    var latitude: Double? = null,
+    var longtitude: Double? = null,
     var id: String? = null,
     var pw: String? = null,
     var phonenum: String? = null
